split plugins to independent projects
