+++ addons/argonaut/src/main/scala/com/github/tminglei/slickpg/PgArgonautSupport.scala
@@ -1,0 +1,55 @@
+package com.github.tminglei.slickpg
+
+import slick.driver.PostgresDriver
+import slick.jdbc.{JdbcType, PositionedResult}
+
+trait PgArgonautSupport extends json.PgJsonExtensions with utils.PgCommonJdbcTypes { driver: PostgresDriver =>
+  import driver.api._
+  import argonaut._, Argonaut._
+
+  def pgjson: String
+
+  /// alias
+  trait JsonImplicits extends ArgonautJsonImplicits
+
+  trait ArgonautJsonImplicits {
+    implicit val argonautJsonTypeMapper =
+      new GenericJdbcType[Json](
+        pgjson,
+        (s) => s.parse.toOption.getOrElse(jNull),
+        (v) => v.nospaces,
+        hasLiteralForm = false
+      )
+
+    implicit def argonautJsonColumnExtensionMethods(c: Rep[Json])(
+      implicit tm: JdbcType[Json], tm1: JdbcType[List[String]]) = {
+        new JsonColumnExtensionMethods[Json, Json](c)
+      }
+    implicit def argonautJsonOptionColumnExtensionMethods(c: Rep[Option[Json]])(
+      implicit tm: JdbcType[Json], tm1: JdbcType[List[String]]) = {
+        new JsonColumnExtensionMethods[Json, Option[Json]](c)
+      }
+  }
+
+  trait ArgonautJsonPlainImplicits {
+    import utils.PlainSQLUtils._
+
+    import scala.reflect.classTag
+
+    if (driver.isInstanceOf[ExPostgresDriver]) {
+      driver.asInstanceOf[ExPostgresDriver].bindPgTypeToScala("json", classTag[Json])
+      driver.asInstanceOf[ExPostgresDriver].bindPgTypeToScala("jsonb", classTag[Json])
+    }
+
+    implicit class PgJsonPositionedResult(r: PositionedResult) {
+      def nextJson() = nextJsonOption().getOrElse(jNull)
+      def nextJsonOption() = r.nextStringOption().flatMap(_.parse.toOption)
+    }
+
+    ///////////////////////////////////////////////////////////
+    implicit val getJson = mkGetResult(_.nextJson())
+    implicit val getJsonOption = mkGetResult(_.nextJsonOption())
+    implicit val setJson = mkSetParameter[Json](pgjson, _.nospaces)
+    implicit val setJsonOption = mkOptionSetParameter[Json](pgjson, _.nospaces)
+  }
+}
