--- src/main/resources/application.conf
@@ -1,16 +1,0 @@
-slick {
-  ansiDump = true
-  unicodeDump = true
-  sqlIndent = true
-  verifyTypes = true
-}
-
-tsql {
-  driver = "com.github.tminglei.slickpg.MyDate2PostgresDriver"
-  db {
-    connectionPool = disabled
-    driver = "org.postgresql.Driver"
-    url = "jdbc:postgresql://172.17.0.1/test?user=postgres"
-    keepAliveConnection = true
-  }
-}
