+++ src/test/scala/com/github/tminglei/slickpg/PgHStoreSupportSuite.scala
@@ -9,7 +9,7 @@ import scala.concurrent.duration._
 class PgHStoreSupportSuite extends FunSuite {
   import MyPostgresDriver.api._
 
-  val db = Database.forURL(url = dbUrl, driver = "org.postgresql.Driver")
+  val db = Database.forURL(url = utils.dbUrl, driver = "org.postgresql.Driver")
 
   case class MapBean(id: Long, hstore: Map[String, String])
