+++ src/test/scala/com/github/tminglei/slickpg/PgLTreeSupportSuite.scala
@@ -10,7 +10,7 @@ import scala.concurrent.duration._
 class PgLTreeSupportSuite extends FunSuite {
   import MyPostgresDriver.api._
 
-  val db = Database.forURL(url = dbUrl, driver = "org.postgresql.Driver")
+  val db = Database.forURL(url = utils.dbUrl, driver = "org.postgresql.Driver")
 
   case class LTreeBean(id: Long, path: LTree, treeArr: List[LTree])
