+++ src/test/scala/com/github/tminglei/slickpg/package.scala
@@ -4,7 +4,5 @@ import java.util.concurrent.Executors
 import scala.concurrent.ExecutionContext
 
 package object slickpg {
-  val dbUrl = "jdbc:postgresql://localhost/test?user=postgres"
-
   implicit val testExecContext = ExecutionContext.fromExecutorService(Executors.newFixedThreadPool(4))
 }
